Move Recorder-Related Files to "src/flash"
