+++ boards/cats_rev1Pro/src/flash/recorder.c
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "util/recorder.h"
+#include "recorder.h"
 #include "util/log.h"
 #include "config/globals.h"
 #include "config/cats_config.h"
