+++ boards/cats_rev1Pro/src/tasks/task_recorder.c
@@ -20,8 +20,8 @@
 #include "tasks/task_recorder.h"
 #include "util/log.h"
 #include "util/types.h"
-#include "lfs/lfs_custom.h"
-#include "util/recorder.h"
+#include "flash/lfs_custom.h"
+#include "flash/recorder.h"
 #include "config/cats_config.h"
 
 #include <stdlib.h>
